refactor: combine into one DAO proposal
