+++ proposals/dao/tip_121c.ts
@@ -17,19 +17,44 @@ const fipNumber = 'tip_121c';
 
 let pcvStatsBefore: PcvStats;
 
-// Amount of FEI sudo() script mints to accounts[0]
+// Amount of FEI sudo() script mints to accounts[0], will be subtracted off
 const AMOUNT_FEI_MINTED_BY_E2E = toBN('10000000000000000000000000'); // 10M
+
+/////////////  Tribe Redeemer config
+// Circulating amount of TRIBE, which is redeemable for underlying PCV assets
+// TODO: Update with final circulating TRIBE figure
+const REDEEM_BASE = ethers.constants.WeiPerEther.mul(458_964_340);
+
+// Lido deposit balance, being withdrawn and sent to Tribe Redeemer
+const STETH_DEPOSIT_BALANCE = '50296523674661485703301';
+const DAO_TIMELOCK_FOX_BALANCE = '15316691965631380244403204';
+const DAO_TIMELOCK_LQTY_BALANCE = '1101298805118942906652299';
+
+// Minimum DAI transferred to Redeemer. Lower bound
+// TODO: Update with final numbers
+const REMAINING_DEPOSIT_DAI_FOR_REDEEMER = ethers.constants.WeiPerEther.mul(30_000_000);
 
 // Do any deployments
 // This should exclusively include new contract deployments
 const deploy: DeployUpgradeFunc = async (deployAddress: string, addresses: NamedAddresses, logging: boolean) => {
+  // SimpleFeiPSM
   const simpleFeiDaiPSMFactory = await ethers.getContractFactory('SimpleFeiDaiPSM');
   const simpleFeiDaiPSM = await simpleFeiDaiPSMFactory.deploy();
   await simpleFeiDaiPSM.deployed();
-  logging && console.log(`simpleFeiDaiPSM: ${simpleFeiDaiPSM.address}`);
+  console.log('Simple FEI PSM deployed to: ', simpleFeiDaiPSM.address);
+
+  // Tribe Redeemer
+  const tribeRedeemerFactory = await ethers.getContractFactory('TribeRedeemer');
+  const redeemedToken = addresses.tribe;
+  const tokensReceived = [addresses.steth, addresses.lqty, addresses.fox, addresses.dai];
+
+  const tribeRedeemer = await tribeRedeemerFactory.deploy(redeemedToken, tokensReceived, REDEEM_BASE);
+  await tribeRedeemer.deployTransaction.wait();
+  console.log('Tribe redeemer deployed to: ', tribeRedeemer.address);
 
   return {
-    simpleFeiDaiPSM
+    simpleFeiDaiPSM,
+    tribeRedeemer
   };
 };
 
@@ -85,9 +110,6 @@ const validate: ValidateUpgradeFunc = async (addresses, oldContracts, contracts,
     '(millions)'
   );
   console.log('simpleFeiDaiPSM FEI balance', (await contracts.fei.balanceOf(addresses.simpleFeiDaiPSM)) / 1e18);
-
-  // 0. Verify PCV equity diff is 0
-  expect(eqDiff).to.equal(0);
 
   // 1. Verify old DAI PSM is deprecated
   // Has no funds
@@ -146,6 +168,26 @@ const validate: ValidateUpgradeFunc = async (addresses, oldContracts, contracts,
   // Burn the 1000 FEI held on the contract
   await contracts.simpleFeiDaiPSM.burnFeiHeld();
   expect(await contracts.fei.balanceOf(addresses.simpleFeiDaiPSM)).to.equal(0);
+
+  //////////////////////    TRIBE REDEEMER    /////////////////////
+  // 4. Verify Tribe Redeemer contract deploy params
+  expect(await contracts.tribeRedeemer.redeemBase()).to.equal(REDEEM_BASE);
+  expect(await contracts.tribeRedeemer.redeemedToken()).to.equal(addresses.tribe);
+
+  const expectedTokensReceived = [addresses.steth, addresses.lqty, addresses.fox, addresses.dai];
+  const actualTokensReceived = await contracts.tribeRedeemer.tokensReceivedOnRedeem();
+  expect(actualTokensReceived.length).to.equal(expectedTokensReceived.length);
+  expect(actualTokensReceived).to.contain(actualTokensReceived[0]);
+  expect(actualTokensReceived).to.contain(actualTokensReceived[1]);
+  expect(actualTokensReceived).to.contain(actualTokensReceived[2]);
+  expect(actualTokensReceived).to.contain(actualTokensReceived[3]);
+
+  // 5. Verify Tribe Redeemer has all PCV assets deposited on it
+  // stETH is rebasing, so amount transfered to tribeRedeemer will be greater than current amount on deposit
+  expect(await contracts.steth.balanceOf(addresses.tribeRedeemer)).to.be.bignumber.greaterThan(STETH_DEPOSIT_BALANCE);
+  expect(await contracts.dai.balanceOf(addresses.tribeRedeemer)).to.equal(REMAINING_DEPOSIT_DAI_FOR_REDEEMER);
+  expect(await contracts.lqty.balanceOf(addresses.tribeRedeemer)).to.equal(DAO_TIMELOCK_LQTY_BALANCE);
+  expect(await contracts.fox.balanceOf(addresses.tribeRedeemer)).to.equal(DAO_TIMELOCK_FOX_BALANCE);
 };
 
 export { deploy, setup, teardown, validate };
