+++ protocol-configuration/proposalsConfig.ts
@@ -2,7 +2,6 @@ import { ProposalCategory, TemplatedProposalsConfigMap } from '@custom-types/typ
 
 import fip_x from '@proposals/description/fip_x';
 import tip_121c from '@proposals/description/tip_121c';
-import tip_121c_pt2 from '@proposals/description/tip_121c_pt2';
 import tip_auraotc from '@proposals/description/tip_auraotc';
 
 export const ProposalsConfig: TemplatedProposalsConfigMap = {
@@ -19,15 +18,6 @@ export const ProposalsConfig: TemplatedProposalsConfigMap = {
     deploy: true, // deploy flag for whether to run deploy action during e2e tests or use mainnet state
     totalValue: 0, // amount of ETH to send to DAO execution
     proposal: tip_121c, // full proposal file, imported from '@proposals/description/fip_xx.ts'
-    proposalId: '',
-    affectedContractSignoff: [],
-    deprecatedContractSignoff: [],
-    category: ProposalCategory.DAO
-  },
-  tip_121c_pt2: {
-    deploy: true, // deploy flag for whether to run deploy action during e2e tests or use mainnet state
-    totalValue: 0, // amount of ETH to send to DAO execution
-    proposal: tip_121c_pt2, // full proposal file, imported from '@proposals/description/fip_xx.ts'
     proposalId: '',
     affectedContractSignoff: [],
     deprecatedContractSignoff: [],
